Input component enhancements (#78)
* adjust input element dimensions
Fixes some style issues related to paddings,
margins, font sizes and line heights. Also uses
the standards we are using in the codebase for
the interplay between styled components and
typescript.

* update tests, stories, readme for input

* fix input focus styles

* fix input border and placeholder
Adds  the missing border for focus  and
active states. Also removes the specific
placholder color for active state

* fix input border colors
Correct border colors were not being applied
in the cases of error and success.

* update tests
undoes a change a changing a span
to a div which was unneeded. Also, removes
border: none from the input as there should
always be a transparent border instead. Also,
updates the snpahots.